Move user signup service to services package
