+++ h/services/__init__.py
@@ -9,3 +9,4 @@ def includeme(config):
     config.register_service_factory('.nipsa.nipsa_factory', name='nipsa')
     config.register_service_factory('.settings.settings_factory', name='settings')
     config.register_service_factory('.user.user_service_factory', name='user')
+    config.register_service_factory('.user_signup.user_signup_service_factory', name='user_signup')
